+++ src/preferences/info.rs
@@ -1,0 +1,8 @@
+use serde::{Deserialize, Serialize};
+
+use crate::puzzle::TwistMetric;
+
+#[derive(Serialize, Deserialize, Debug, Default, Clone)]
+pub struct InfoPreferences {
+    pub metric: TwistMetric,
+}