+++ src/puzzle/generic.rs
@@ -3,7 +3,9 @@ use std::fmt;
 use std::time::Duration;
 use thiserror::Error;
 
-use super::{rubiks3d, rubiks4d, traits::*, PuzzleController, PuzzleType, Rubiks3D, Rubiks4D};
+use super::{
+    rubiks3d, rubiks4d, traits::*, PuzzleController, PuzzleType, Rubiks3D, Rubiks4D, TwistMetric,
+};
 use crate::render::WireframeVertex;
 
 /// A PuzzleController of any puzzle type.